feat (order) : orderMapper, orderValidator 삭제
- 주문 완료로 상태 변경 시 OrderCompletionEvent 를 publish 해 table 의 empty 상태를 바꿈
- readme 에 객체 의존 관계 이미지 추가
